Move timeshift settings overlay to correct namespace
